extract PcntlFork
